+++ driven/groups/adapter.go
@@ -94,17 +94,18 @@ func (a *Adapter) GetGroupsMembership(userToken string) (*GroupMembership, error
 }
 
 // GetGroupDetails retrieves group details
-func (a *Adapter) GetGroupDetails(groupID string) (*model.Group, error) {
+func (a *Adapter) GetGroupDetails(userToken string, groupID string) (*model.Group, error) {
 	if groupID != "" {
 
-		url := fmt.Sprintf("%s/api/int/group/%s", a.baseURL, groupID)
+		url := fmt.Sprintf("%s/api/v2/groups/%s", a.baseURL, groupID)
 		client := &http.Client{}
 		req, err := http.NewRequest("GET", url, nil)
-		req.Header.Set("INTERNAL-API-KEY", a.internalAPIKey)
 		if err != nil {
 			log.Printf("error GetGroupDetails: request - %s", err)
 			return nil, fmt.Errorf("error GetGroupDetails: request - %s", err)
 		}
+
+		req.Header.Add("Authorization", fmt.Sprintf("Bearer %s", userToken))
 
 		resp, err := client.Do(req)
 		if err != nil {
@@ -175,4 +176,40 @@ func (a *Adapter) sendGroupNotification(groupID string, notification model.Group
 			log.Printf("error SendGroupNotification: request - %d. Error: %s", resp.StatusCode, err)
 		}
 	}
+}
+
+// UpdateGroupDateUpdated Updates group date updated
+func (a *Adapter) UpdateGroupDateUpdated(groupID string) error {
+	if groupID != "" {
+
+		url := fmt.Sprintf("%s/api/int/group/%s/date_updated", a.baseURL, groupID)
+		client := &http.Client{}
+		req, err := http.NewRequest("POST", url, nil)
+		if err != nil {
+			log.Printf("error UpdateGroupDateUpdated: request - %s", err)
+			return fmt.Errorf("error UpdateGroupDateUpdated: request - %s", err)
+		}
+
+		req.Header.Add("INTERNAL-API-KEY", a.internalAPIKey)
+
+		resp, err := client.Do(req)
+		if err != nil {
+			log.Printf("error UpdateGroupDateUpdated: request - %s", err)
+			return fmt.Errorf("error UpdateGroupDateUpdated: request - %s", err)
+		}
+		defer resp.Body.Close()
+		if resp.StatusCode != 200 {
+			errorBody, err := ioutil.ReadAll(resp.Body)
+			if err != nil {
+				log.Printf("error UpdateGroupDateUpdated: request - %s", err)
+				return fmt.Errorf("error UpdateGroupDateUpdated: request - %s", err)
+			}
+
+			log.Printf("error UpdateGroupDateUpdated: request - %d. Error: %s, Body: %s", resp.StatusCode, err, string(errorBody))
+			return fmt.Errorf("error UpdateGroupDateUpdated: request - %d. Error: %s, Body: %s", resp.StatusCode, err, string(errorBody))
+		}
+
+		return nil
+	}
+	return nil
 }